Adding bar char to components
